Refactored transformations
Split up the abstract base class for transformations and actual transformation classes.
